backend/requirements.txt
fastapi==0.115.6
pydantic-settings==2.7.0
SQLAlchemy==2.0.36
uvicorn[standard]==0.34.0
